backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
supabase>=2.5.0
anthropic>=0.28.0
openai>=1.30.0
python-docx>=1.1.2
pypdf>=4.2.0
httpx>=0.27.0
pydantic>=2.7.1
python-multipart>=0.0.9
